Refactored TimelineCreateButtonGroup
