+++ packages/twenty-front/src/modules/activities/timeline/components/Timeline.tsx
@@ -1,15 +1,8 @@
 import styled from '@emotion/styled';
-import { useSetRecoilState } from 'recoil';
-import { Button, ButtonGroup } from 'tsup.ui.index';
 
-import { useOpenCreateActivityDrawer } from '@/activities/hooks/useOpenCreateActivityDrawer';
+import { TimelineCreateButtonGroup } from '@/activities/timeline/components/TimelineCreateButtonGroup';
 import { useTimelineActivities } from '@/activities/timeline/hooks/useTimelineActivities';
 import { ActivityTargetableObject } from '@/activities/types/ActivityTargetableEntity';
-import {
-  IconCheckbox,
-  IconNotes,
-  IconPaperclip,
-} from '@/ui/display/icon/index';
 import AnimatedPlaceholder from '@/ui/layout/animated-placeholder/components/AnimatedPlaceholder';
 import {
   StyledEmptyContainer,
@@ -17,8 +10,6 @@ import {
   StyledEmptyTextContainer,
   StyledEmptyTitle,
 } from '@/ui/layout/animated-placeholder/components/EmptyPlaceholderStyled';
-import { TAB_LIST_COMPONENT_ID } from '@/ui/layout/show-page/components/ShowPageRightContainer';
-import { useTabList } from '@/ui/layout/tab/hooks/useTabList';
 import { useIsMobile } from '@/ui/utilities/responsive/hooks/useIsMobile';
 
 import { TimelineItemsContainer } from './TimelineItemsContainer';
@@ -40,14 +31,9 @@ export const Timeline = ({
 }: {
   targetableObject: ActivityTargetableObject;
 }) => {
-  const { getActiveTabIdState } = useTabList(TAB_LIST_COMPONENT_ID);
-  const setActiveTabId = useSetRecoilState(getActiveTabIdState());
-
   const { activities, initialized } = useTimelineActivities({
     targetableObject,
   });
-
-  const openCreateActivity = useOpenCreateActivityDrawer();
 
   const showEmptyState = initialized && activities.length === 0;
 
@@ -68,33 +54,7 @@ export const Timeline = ({
             There are no activities associated with this record.{' '}
           </StyledEmptySubTitle>
         </StyledEmptyTextContainer>
-        <ButtonGroup variant={'secondary'}>
-          <Button
-            Icon={IconNotes}
-            title="Note"
-            onClick={() =>
-              openCreateActivity({
-                type: 'Note',
-                targetableObjects: [targetableObject],
-              })
-            }
-          />
-          <Button
-            Icon={IconCheckbox}
-            title="Task"
-            onClick={() =>
-              openCreateActivity({
-                type: 'Task',
-                targetableObjects: [targetableObject],
-              })
-            }
-          />
-          <Button
-            Icon={IconPaperclip}
-            title="File"
-            onClick={() => setActiveTabId('files')}
-          />
-        </ButtonGroup>
+        <TimelineCreateButtonGroup targetableObject={targetableObject} />
       </StyledEmptyContainer>
     );
   }
